fix: fix tests
