+++ src/components/CardForm.tsx
@@ -14,7 +14,9 @@ const CardForm: React.FC<CardProps> = ({ title, person, date, imageCard, personI
   return (
     <div className="card">
       <CardImage imageCard={imageCard} />
-      <div className="card_title">{title}</div>
+      <div className="card_title" role="title">
+        {title}
+      </div>
       <CardInfo person={person} personImg={personImg} date={date} />
     </div>
   );
